move react components into separate files
