+++ src/ui/Highlight.less
@@ -1,0 +1,5 @@
+.Highlight {
+  .highlighted {
+    background-color: yellow;
+  }
+}
